requirements.txt
requests
flask
google-cloud-bigquery
google-cloud-storage